+++ src/commands/ban.ts
@@ -1,5 +1,5 @@
 import Command from '../interfaces/command';
-import { Message, Collection } from 'discord.js';
+import { Message, Collection, GuildMember } from 'discord.js';
 
 export default class BanCommand implements Command {
     info = {
@@ -8,7 +8,23 @@ export default class BanCommand implements Command {
         usage: '&ban'
     }
 
-    async run(message: Message, args: string[], messages: any): Promise<void> {
-        
+    run(message: Message, args: string[], messages: any): any {
+        if(!message.member.hasPermission('BAN_MEMBERS'))
+            return message.reply(messages.noPermission);
+
+        let member: GuildMember = message.mentions.members.first()
+            || message.guild.members.find(
+                (m: GuildMember): boolean => m.displayName == args.join(' ')
+            );
+
+        if(!member)
+            return message.reply(messages.couldNotFind);
+        if(!member.bannable)
+            return message.reply(messages.couldNotBan);
+
+        member.ban().then((member: GuildMember): void => {
+            member.user.send(messages.youWereBanned.replace('{}', message.guild.name));
+            message.reply(messages.banned.replace('{}', member.user.username));
+        });
     }
 }